Modify files: Delete: - TfmBootstrap - TfmRestAPI Add: -  TfmApp
New features:
Add homeworkDone list to homework, to get the homework done for the students more easy. (Teacher create homework with description, available and date. Student modify only the file_link field on the homework. Teacher finally modify the grade and coment the homework.)

Deleted features:
The behavior of the student with the homework change to: Homework have an instance of student. Instead of student have a list of homework.